+++ src/blocks/StartBlock.cpp
@@ -1,0 +1,6 @@
+#include "StartBlock.h"
+
+StartBlock::StartBlock(QWidget *parent)
+    : Block("Start", QColor("#2d7a2d"), parent)
+{
+}