+++ src/blocks/StartBlock.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Block.h"
+
+class StartBlock : public Block {
+    Q_OBJECT
+
+public:
+    explicit StartBlock(QWidget *parent = nullptr);
+};