+++ src/blocks/StopBlock.cpp
@@ -1,0 +1,6 @@
+#include "StopBlock.h"
+
+StopBlock::StopBlock(QWidget *parent)
+    : Block("Stop", QColor("#7a2d2d"), parent)
+{
+}