+++ src/blocks/StopBlock.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Block.h"
+
+class StopBlock : public Block {
+    Q_OBJECT
+
+public:
+    explicit StopBlock(QWidget *parent = nullptr);
+};